Changes to meet requirements
Created a function, setTextFieldAttributes, that takes a text field as
an argument and sets some attributes for it.
Changed keyboardWillShow to only shift view if the bottomTextField is
selected.
Created a function, pickAnImage, for image picking to handle the
redundancies between pickAnImageFromAlbum and pickAnImageFromCamera.
Removed the meme array creation from shareMeme and implemented
completionWithItemsHandler to save the meme once it was shared.
Removed unneeded uses of self from generateMemedImage().